requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.2
Flask-Admin==1.6.1
Flask-BasicAuth==0.2.0
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
prompt-toolkit==3.0.43
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.2
regex==2023.12.25
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.2
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.2
